Seperating lookup and update into their own actors, now working test that Updater calls out to Lookup in order to request the item ID
